+++ src/ar/com/untref/imagenes/bordes/DetectorDeBordesLorentz.java
@@ -10,7 +10,7 @@ public class DetectorDeBordesLorentz implements InterfaceDetectorDeBordes{
 	
 	@Override
 	public float gradiente(float derivada) {
-		return (float) (1/((Math.pow(Math.abs(derivada), 2) / Math.pow(sigma, 2)) + 1));
+		return (float) ( 1/ ( ((float) (Math.pow(Math.abs(derivada), 2) / Math.pow(sigma, 2))) + 1) );
 	}
 	
 }
